[perf] cut three.js from the initial bundle and fix first paint
