+++ public/index.html
@@ -3,6 +3,13 @@
   <head>
     <meta charset="utf-8" />
 
+    <!-- Night is always the first frame; don't flash white while React boots -->
+    <style>
+      html {
+        background: #000;
+      }
+    </style>
+
     <!-- Primary Meta Tags -->
     <title>Andrew Hunt - Frontend Engineer | New York</title>
     <meta
