+++ src/index.css
@@ -9,12 +9,6 @@
 @config "../tailwind.config.js";
 
 @font-face {
-  font-family: "infini";
-  src: url("./infini-regular.woff");
-  font-display: swap;
-}
-
-@font-face {
   font-family: "nova-mono";
   src: url("./NovaMono-Regular.ttf");
   font-display: swap;
@@ -31,8 +25,8 @@ body {
   margin: 0;
   padding: 0;
   font-family:
-    -apple-system, "infini", "Helvetica Neue", "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", sans-serif;
+    -apple-system, "Helvetica Neue", "Segoe UI", "Roboto", "Oxygen", "Ubuntu",
+    "Cantarell", "Fira Sans", "Droid Sans", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
